+++ Module/Rule/Functions.ps1
@@ -146,7 +146,9 @@ function Get-HardCodedString
         'V-8322.b', # Time Synchronization
         'V-14235', # UAC - Admin Elevation Prompt
         'V-26359', # Windows Server - Legal Banner Dialog Box Title
-        'V-17761' # Outlook 2013 - OrgSetting Value
+        'V-17761', # Outlook 2013 - OrgSetting Value
+        'V-75241', # Windows Defender - ASSignatureDue
+        'V-75243' # Windows Defender - AVSignatureDue
     )
 
     if ($stigIds -contains $stigId)
@@ -205,6 +207,10 @@ function Get-HardCodedString
         {
             $hardCodedString = "'{0}' -ge '30' -and '{0}' -le '132'"
             continue
+        }
+        {$PSItem -match 'V-75241|V-75243'}
+        {
+            $hardCodedString = "{0} -ge '1' -and {0} -le '7'"
         }
     }