+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   * (DC only) V-2376 - V-2380 are migrated from manual to account policy rules.
 * Added the following STIGs
   * SQL Server 2016 Instance V1R3 [#186](https://github.com/Microsoft/PowerStig/issues/186): Update PowerSTIG to parse the SQL Server 2016 Instance STIG - V1R3
+  * Windows Defender Antivirus V1R4 [#236](https://github.com/microsoft/PowerStig/issues): Update PowerSTIG to parse the Windows Defender STIG with new Registry Rule
 
 ## 2.3.2.0